progress: savepoint: seems the Path object can't be used.
I've to figure out how to save the touchpoints and render them.
I've figured out that imperative is re-rendered based on react state updates - not sure yet if the blendmode is been applied - but it at least is rendering.

So, once I figure out how to get the touch points there - that should be definitive progress.